finaliza construção do aplicativo de gestao financeira!

+++ Models/Transaction.cs
@@ -14,7 +14,7 @@ namespace AppControleFinanceiro.Models
         public TransactionType Type { get; set; }
         public string Name { get; set; } = string.Empty;
         public DateTimeOffset Date { get; set; }
-        public decimal Value { get; set; }
+        public double Value { get; set; }
     }
 
     public enum TransactionType
